Parsing tags, mentions (users) from entries
Tags and mention tokens are words starting with '#' and '@'
respectively. Using group-by to partition the tokens into two lists -
words starting with and those without the 2 characters respectively.

Since tags, mentions and balance essentially extract the matches & the
non-matches from tokens and each takes tokens as arguments created a pipelining
function '|>>' which essentially pipelines those methods which take
tokens as arguments and in turn returns a list containing matches and a
list with non-matches.

|>> can be applied to any set of homogenic functions that take 1 arguments
but returns a 2 tuple - where the first result is used but the second one
passed on.